+++ fighters/demon/src/status/attack_lw3.rs
