+++ fighters/demon/src/status/attack_stand_2.rs
